refactor: AddFriendBottomSheet를 책임 분리 (SRP 적용)
